+++ apps/trackers/jira/query.py
@@ -142,7 +142,6 @@ IMPACT_TO_JIRA_CVE_SEVERITY = {
 MAJOR_INCIDENT = "Major Incident"
 KEV = "KEV (active exploit case)"
 COMPLIANCE_PRIORITY = "Compliance Priority"
-CONTRACT_PRIORITY = "Contract Priority"
 
 
 class OldTrackerJiraQueryBuilder(TrackerQueryBuilder):
@@ -344,10 +343,6 @@ class OldTrackerJiraQueryBuilder(TrackerQueryBuilder):
             Affect.AffectAffectedness.NEW
         }:
             self._query["fields"]["labels"].append("validation-requested")
-
-        # If at least one affect has is_contract_priority, add label contract-priority
-        if self.tracker.is_contract_priority:
-            self._query["fields"]["labels"].append("contract-priority")
 
         if self.tracker.is_compliance_priority:
             self._query["fields"]["labels"].append("compliance-priority")
@@ -778,7 +773,6 @@ class TrackerJiraQueryBuilder(OldTrackerJiraQueryBuilder):
             MAJOR_INCIDENT,
             KEV,
             COMPLIANCE_PRIORITY,
-            CONTRACT_PRIORITY,
         ]
         missing_allowed_values = sorted(
             set(expected_allowed_values) - set(allowed_values)
@@ -794,7 +788,6 @@ class TrackerJiraQueryBuilder(OldTrackerJiraQueryBuilder):
         choice_major_incident = False
         choice_kev = False
         choice_compliance_priority = False
-        choice_contract_priority = False
 
         flaw_uuids = self.tracker.affects.values_list("flaw__uuid", flat=True)
 
@@ -807,7 +800,6 @@ class TrackerJiraQueryBuilder(OldTrackerJiraQueryBuilder):
                 flaw.major_incident_state == Flaw.FlawMajorIncident.CISA_APPROVED
             )
             choice_compliance_priority |= self.tracker.is_compliance_priority
-            choice_contract_priority |= self.tracker.is_contract_priority
 
         multichoice_jira_field_value = []
         if choice_major_incident:
@@ -816,8 +808,6 @@ class TrackerJiraQueryBuilder(OldTrackerJiraQueryBuilder):
             multichoice_jira_field_value.append({"value": KEV})
         if choice_compliance_priority:
             multichoice_jira_field_value.append({"value": COMPLIANCE_PRIORITY})
-        if choice_contract_priority:
-            multichoice_jira_field_value.append({"value": CONTRACT_PRIORITY})
 
         self._query["fields"][field_id] = multichoice_jira_field_value