+++ collectors/ps_constants/tests/test_core.py
@@ -4,17 +4,11 @@ from apps.trackers.models import JiraBugIssuetype
 from collectors.ps_constants.core import (
     fetch_ps_constants,
     sync_compliance_priority,
-    sync_contract_priority,
     sync_jira_bug_issuetype,
     sync_special_consideration_packages,
     sync_ubi_packages,
 )
-from osidb.models import (
-    CompliancePriority,
-    ContractPriority,
-    SpecialConsiderationPackage,
-    UbiPackage,
-)
+from osidb.models import CompliancePriority, SpecialConsiderationPackage, UbiPackage
 
 pytestmark = pytest.mark.unit
 
@@ -71,11 +65,6 @@ class TestPsConstantsCollection:
             "rhel-9",
         ]
 
-        url = "/".join((ps_constant_base_url, "contract_priority.yml"))
-        contract_priority = fetch_ps_constants(url)
-        assert "rhel-8" in contract_priority
-        assert len(contract_priority["rhel-8"]) > 0
-
         url = "/".join((ps_constant_base_url, "ubi_packages.yml"))
         ubi_packages = fetch_ps_constants(url)
         assert "rhel-9" in ubi_packages
@@ -120,17 +109,6 @@ class TestPsConstantsCollection:
             "cluster-monitoring-operator-container",
         ]
         assert CompliancePriority.objects.get(ps_module="foobar-123").components == []
-
-    def test_sync_contract_priority(self):
-        """
-        test collector contract priority data sync
-        """
-        # The meaning of SAMPLE_DATA here is {"module": ["update stream"]}.
-        sync_contract_priority(SAMPLE_DATA)
-
-        assert ContractPriority.objects.all().count() == 4
-        assert ContractPriority.objects.filter(ps_update_stream="test1").count() == 1
-        assert ContractPriority.objects.filter(ps_update_stream="test2").count() == 1
 
     def test_sync_ubi_packages(self):
         """Check collector can correctly sync ubi data in database"""