+++ osidb/__init__.py
@@ -1,5 +1,5 @@
 """
     osidb version
 """
-__version__ = "4.3.0"
+__version__ = "4.3.1"
 VERSION = __version__