+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [4.3.1] - 2024-09-11
 ### Added
 - Create new API endpoints for DjangoQL (OSIDB-3338)
 - Implement Jira collector sync managers (OSIDB-3177)